# Dockerfile: optimize for caching