@thaJeztah 2019-03-06
@L-Hudson PTAL; from local testing, the archives all work as before, but pls double-check if there are corner-cases that I overlooked.