@bermudezmt 2019-03-07
@thaJeztah I got done in 98.192 seconds while testing this locally, which is even better than your experience. It does look like the build "hangs" on step 30, but I'm wondering if this will go away once I fix the Liquid syntax errors which I've introduced from a previous PR. The archives do work. :)