@bermudezmt 2019-03-07
```
real    3m7.848s
user    0m2.911s
sys     0m1.087s
```
What I got on my end with BUILDKIT enabled. Thank you, @thaJeztah !